More warning fixes
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@238 37977431-3df6-0310-b722-df95706aa16b

+++ rts/AirMoveType.cpp
@@ -759,7 +759,7 @@ void CAirMoveType::UpdateAirPhysics(float rudder, float aileron, float elevator,
 #ifdef DIRECT_CONTROL_ALLOWED
 	if(owner->directControl)
 		if((pos.y-gHeight)>wantedHeight*1.2)
-			engine=max(0,min(engine,1-(pos.y-gHeight-wantedHeight*1.2)/wantedHeight));
+			engine=max(0.,min((double)engine,1-(pos.y-gHeight-wantedHeight*1.2)/wantedHeight));
 #endif
 
 	speed+=engineVector*maxAcc*engine;
